+++ R/library.R
@@ -1,3 +1,18 @@
+
+fileName <- function(mf,db,organism,path = 'construction_library'){
+  
+  if (length(organism) == 0) {
+    organism <- 'none'
+  }
+  
+  file_name <- switch(db,
+                      kegg = paste0(path,'/',mf,'_',db,'_',organism,'.rds'),
+                      pubchem = paste0(path,'/',mf,'_',db,'.rds')
+  )
+  
+  return(file_name)
+}
+
 #' @importFrom readr write_rds
 
 setGeneric('saveConsensus',function(x,path = 'construction_library')
@@ -6,87 +21,80 @@ setGeneric('saveConsensus',function(x,path = 'construction_library')
 setMethod('saveConsensus',signature = 'Consensus',
           function(x,path = 'construction_library'){
             
-            message(str_c('Exporting to ',path))
-            
-            db <- database(x)
-            
-            if (length(organism(x)) == 0) {
-              org <- 'none'
-            } else {
-              org <- organism(x)
+            if (!dir.exists(path)){
+              dir.create(path,recursive = TRUE)
             }
             
-            fileName <- switch(db,
-                               kegg = str_c(str_c(mf(x),db,org,sep = '_'),'.rds'),
-                               pubchem = str_c(str_c(mf(x),db,sep = '_'),'.rds'))
+            message(str_c('Exporting to ',path))
             
-            write_rds(x,str_c(path,fileName,sep = '/'))
+            write_rds(
+              x,
+              fileName(
+                mf(x),
+                database(x),
+                organism(x),
+                path = path
+              ))
           })
 
 #' @importFrom readr read_rds
 
-loadLibrary <- function(MFs, path = 'construction_library'){
+status <- function(items,library_path,threshold = 50){
   
-  if (!dir.exists(path)) {
-    dir.create(
-      path,
-      recursive = TRUE)  
-    message(str_c('Created construction library at ',path))
-  } else {
-    message(str_c('Using construction library at ',path))
-  }
+  item_status <- items %>% 
+    rowwise() %>% 
+    mutate(
+      file_name = fileName(MF,database,organism,path = library_path),
+      exists = file.exists(file_name),
+      status = if (exists) {
+        read_rds(file_name) %>% 
+          consensus(Adduct,threshold) %>% 
+          .$kingdom %>% 
+          .[1]
+      } else {
+        'Unavailable'
+      },
+      status = status %>%
+        replace(
+          !. %in% c(
+            'No database hits',
+            'Unclassified',
+            'Unavailable'
+          ),
+          'Classified'
+        ),
+      status = factor(status,
+                      levels = c(
+                        'Classified',
+                        'Unclassified',
+                        'No database hits',
+                        'Unavailable'
+                      ))
+    )
   
-  libraryContents <- list.files(path,full.names = TRUE,pattern = '.rds') %>%
-    tibble(MF = basename(.) %>%
-             str_split_fixed('_',2) %>%
-             .[,1],
-           path = .) %>%
-    filter(MF %in% MFs$MF)
-  
-  libraryContents %>%
-    .$path %>%
-    map(read_rds)
+  return(item_status)
 }
 
-setGeneric('status',function(x){
-  standardGeneric('status')
-})
 
-setMethod('status',signature = 'Consensus',
-          function(x){
-            
-            if (length(organism(x)) == 0) {
-              org <- 'none'
-            } else {
-              org <- organism(x)
-            }
-            
-            st <- x %>%
-              consensusClassifications() %>%
-              .$kingdom %>%
-              unique() %>%
-              sort()
-            
-            if (length(st) == 1 & 
-                !identical(st,'No database hits') & 
-                !identical(st,'Unclassified')) {
-              st <- 'Classified'
-            }
-            
-            if (length(st) > 1) {
-              
-              if ('Unclassified' %in% st) {
-                if (identical(st,c('No database hits','Unclassified'))) {
-                  st <- 'Unclassified'
-                } else {
-                  st <- 'Classified'
-                }  
-              } else {
-                st <- 'Classified'
-              }
-            }
-            
-            
-            tibble(MF = mf(x),organism = org,database = database(x),status = st)
-          }
-)
+searchNecessary <- function(search_list,db){
+  
+  search_list <- search_list %>% 
+    dplyr::filter(
+      status != 'Classified'
+    )
+  
+  if ('kegg' %in% db){
+    search_list <- search_list %>% 
+      dplyr::group_by(MF) %>% 
+      dplyr::filter(
+        'kegg' %in% database
+      )
+  }
+  
+  search_list <- search_list %>% 
+    filter(
+      status == 'Unavailable'
+    )
+  
+  return(search_list)
+}