+++ R/structural_classifications.R
@@ -1,0 +1,7 @@
+#' Example structural classifications
+#' @format A tibble containing consensus structural classifications for example *m/z* features.
+#' @source These example consensus structural classifications are based on the *Brachypodium distachyon* ecotype 
+#' comparison FIE-MS metabolomic fingerprinting data set available in the 
+#' [{metaboData}](https://aberhrml.github.io/metaboData/) package.
+
+"structural_classifications"